create player hand
GH-3
